requirements.txt
# TravelMate AI - Python Dependencies
# Install these packages using: pip install -r requirements.txt

# FastAPI - Modern web framework for building APIs
fastapi

# Uvicorn - ASGI server for running FastAPI applications
uvicorn[standard]

# Pydantic - Data validation (included with FastAPI, but listed for clarity)
pydantic

# OpenAI - For AI-powered itinerary generation
openai

# Python-dotenv - For managing environment variables
python-dotenv
